requirements.txt
Flask==0.10.1
Frozen-Flask==0.11
itsdangerous==0.24
Jinja2==2.7.3
Markdown==2.6.2
MarkupSafe==0.23
PyYAML==3.11
Werkzeug==0.10.4
wheel==0.24.0
